+++ dsr_agents/include/dsr_agents/agents/person_agent.hpp
@@ -1,6 +1,7 @@
 /*
- * TF AGENT ROS NODE
+ * PERSON AGENT ROS NODE
  *
+ * Copyright (c) 2023 Óscar Pons Fernández <ajtudela@gmail.com>
  * Copyright (c) 2023 Alberto José Tudela Roldán <ajtudela@gmail.com>
  * 
  * This file is part of dsr_agents.
@@ -9,8 +10,8 @@
  *
  */
 
-#ifndef DSR_AGENT__TF_AGENT_HPP_
-#define DSR_AGENT__TF_AGENT_HPP_
+#ifndef DSR_AGENT__PERSON_AGENT_HPP_
+#define DSR_AGENT__PERSON_AGENT_HPP_
 
 // C++
 #include <string>
@@ -21,7 +22,8 @@
 // ROS
 #include "rclcpp/rclcpp.hpp"
 #include "vision_msgs/msg/detection3_d_array.hpp"
-
+#include "tf2_ros/buffer.h"
+#include "tf2_ros/transform_listener.h"
 // DSR
 #include "dsr/api/dsr_api.h"
 #include "dsr_agents/agent_node.hpp"
@@ -34,6 +36,12 @@ class personAgent: public AgentNode{
 		rclcpp::Subscription<vision_msgs::msg::Detection3DArray>::SharedPtr person_sub_;
 		std::string ros_topic_;
 
+		/// The buffer of the transformations tree.
+		std::unique_ptr<tf2_ros::Buffer> tf_buffer_;
+
+		/// The listener of the transformations tree.
+		std::shared_ptr<tf2_ros::TransformListener> tf_listener_;
+		
 		void get_params();
 		void person_callback(const vision_msgs::msg::Detection3DArray::SharedPtr msg);
 };