+++ dsr_agents/CHANGELOG.rst
@@ -4,8 +4,9 @@ Changelog for package dsr_agents
 
 0.5.X (UNRELEASED)
 ------------------
-* Added function to get priorityof nodes.
+* Added function to get priority of nodes.
 * Fix get_zones in navigation node.
+* Battery agent will publish battery_percentage between 0 and 100.
 
 0.5.0 (07-11-2023)
 ------------------